docs: streamline readme (#5)
## Proposed Changes

Streamline the readme by reducing redundancy and with a more concise
style.

## Types of Changes

What types of changes does your contribution introduce? _Put an `x` in
the boxes
that apply_

- [ ] A **bug fix** (non-breaking change which fixes an issue). Use MR
tag
      `bugfix`.
- [ ] A new **feature** (non-breaking change which adds functionality).
Use MR
      tag `feature`.
- [ ] A **breaking change** (fix or feature that would cause existing
      functionality to not work as expected). Use MR tag `feature`.
- [x] A **non-productive** update (documentation, tooling, etc. if none
of the
      other choices apply). Use MR tag `chore`.

## Checklist

_Put an `x` in the boxes that apply. You can also fill these out after
creating
the PR. If you're unsure about any of them, don't hesitate to ask. We're
here to
help! This is simply a reminder of what we are going to look for before
merging
your code._

- [x] I have read the

[CONTRIBUTING](https://github.com/sdsc-ordes/repository-template/tree/main/CONTRIBUTING.md)
      guidelines.

## Further Comments

<!-- If this is a relatively large or complex change, kick off the
discussion by -->
<!-- explaining why you chose the solution you did and what alternatives
you -->
<!-- considered, etc... -->